(#8662) Fix Puppet.features.root? on Windows
On Windows using the default ruby install, Process.uid always returns 0,
and this was being used to determine if the user was root or not. When
using the win32/processs ruby gem, Process.uid does return a non-zero
value (corresponding to the last part of the SID). But unfortunately
it's Windows, so it's not that simple.

This commit changes Puppet.features.root? and
Puppet::Util::SUIDManager.root? on Windows to return true if the user is
running with elevated privileges (on Vista or 2008) or is a member of
the builtin Administrators group.

It also re-orders the feature definitions, since feature root depends on
the method Puppet::Util::SUIDManager.root?, which depends on
microsoft_windows, which depends on posix, which depends on syslog.