(#8662) Skip user and group resources when applying settings
When running as root, puppet will generate a catalog from its settings
to create the various directories, e.g. var, ssl. If mkusers is true and
a setting implements owner and/or group methods, then puppet will
automatically add user and group resources to the catalog (provided they
are not root or wheel). This functionality will not be supported on
Windows, and so this step is skipped.