(#11944) Warn when running hostname -f on Windows agents
Neither Windows nor cygwin support `hostname -f`, so the acceptance
tests that use it fail on Windows.

This commit changes two of the acceptance tests to log a warning when
run on a Windows agent and skip to the next agent in the list. I've
filed a ticket to re-enable these tests on Windows (#12407).