(#11944) Refactor execution of echo in acceptance tests
Previously, some of the acceptance tests were using `/bin/echo ...` as
the command for an exec resource, which doesn't work on Windows
agents.

This commit refactors the tests to use the `Host#echo` method, which
generates a platform-specific command for executing echo. By default,
the method generates a fully-qualified path for executing echo, but if
you pass false, then it generates an unqualified path (for cases where
the manifest specifies the path to use).

The commit also refactors the tests to use `agents.each do |agent|`
style test, which is necessary to generate platform-specific commands.