Maint: Stub Puppet::Util::Execution.execute method
Recently the execute methods in Puppet::Util were refactored into
Puppet::Util::Execution, but the spec/unit/type/exec_spec.rb test was
still stubbing the old module method. As a result, the test was
executing `/bin/sleep 1`, which does the right thing on Unix. But on
Windows, we get into a situation where we're trying to execute a file
that doesn't exist, resulting in an unexpected error.

This commit just stubs the method on the new module.

Reviewed-by: Matt Robinson <matt@puppetlabs.com>