(#9595) Remove autoflush option
Previously, puppet had an autoflush setting, which if enabled, would
cause the file log destination to call `IO#flush` after writing each
log message. However, this setting was false by default, so if the
agent terminated unexpectedly, any data buffered in ruby, but not yet
flushed, would be lost.

This commit changes the file log destination to not buffer log
messages that it writes. This has the same effect as writing the
message and then calling `IO#flush`, but is more direct.